Re-organize files.
