--- App.config
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<appSettings>
-  
-</appSettings>
