--- App.xaml
@@ -1,6 +1,0 @@
-﻿<Application x:Class="SimpleMvvmSamples.WPF.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="MainWindow.xaml"> 
-</Application>
-
