--- MainWindow.xaml
@@ -1,15 +1,0 @@
-﻿<Window x:Class="SimpleMvvmSamples.WPF.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SimpleMvvmSamples.WPF"
-        xmlns:views="clr-namespace:SimpleMvvmSamples.WPF.Views"
-        mc:Ignorable="d"
-        Title="Main Window" Height="300" Width="525"
-       >
-    <StackPanel>
-        <TextBlock Text="{Binding Path=BannerText}" FontFamily="Verdana" FontSize="18" FontWeight="Bold" HorizontalAlignment="Center" Margin="0,0,0,10" />
-        <!--<views:CustomerView />-->
-    </StackPanel>
-</Window>
