+++ SimpleMvvmToolkit.Samples.WPF.Core.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
-VisualStudioVersion = 16.0.29318.209
+VisualStudioVersion = 16.0.29519.87
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleMvvmToolkit.Samples.WPF.Core", "SimpleMvvmToolkit.Samples.WPF.Core.csproj", "{3A8DEF4F-A444-4B7D-90E8-6F747DC27DFF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleMvvmToolkit.Samples.WPF.Core", "SimpleMvvmToolkit.Samples.WPF.Core\SimpleMvvmToolkit.Samples.WPF.Core.csproj", "{3AE6ED46-7969-4360-9B9F-4B966080CFDA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3A8DEF4F-A444-4B7D-90E8-6F747DC27DFF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3A8DEF4F-A444-4B7D-90E8-6F747DC27DFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3A8DEF4F-A444-4B7D-90E8-6F747DC27DFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3A8DEF4F-A444-4B7D-90E8-6F747DC27DFF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3AE6ED46-7969-4360-9B9F-4B966080CFDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3AE6ED46-7969-4360-9B9F-4B966080CFDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3AE6ED46-7969-4360-9B9F-4B966080CFDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3AE6ED46-7969-4360-9B9F-4B966080CFDA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B4A9A963-4790-45C0-A5A8-6ECE2D831A59}
+		SolutionGuid = {27CB6185-C92B-4D0A-9BB3-2F03C6CBF80A}
 	EndGlobalSection
 EndGlobal
