+++ SimpleMvvmToolkit.Samples.WPF.Core/App.xaml
@@ -1,0 +1,9 @@
+﻿<Application x:Class="SimpleMvvmToolkit.Samples.WPF.Core.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:SimpleMvvmToolkit.Samples.WPF.Core"
+             StartupUri="MainWindow.xaml">
+    <Application.Resources>
+         
+    </Application.Resources>
+</Application>
