+++ SimpleMvvmToolkit.Samples.WPF.Core/MainWindow.xaml
@@ -1,0 +1,13 @@
+﻿<Window x:Class="SimpleMvvmToolkit.Samples.WPF.Core.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:SimpleMvvmToolkit.Samples.WPF.Core"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+    <StackPanel>
+        <TextBlock Text="{Binding Path=BannerText}" FontFamily="Verdana" FontSize="18" FontWeight="Bold" HorizontalAlignment="Center" Margin="0,0,0,10" />
+        <!--<views:CustomerView />-->
+    </StackPanel>
+</Window>
