+++ SimpleMvvmToolkit.Samples.WPF.Core/SimpleMvvmToolkit.Samples.WPF.Core.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <UseWPF>true</UseWPF>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="SimpleMvvmToolkit.Express" Version="2.0.0" />
+  </ItemGroup>
+
+</Project>
